feat: ErrorOp, ErrorTxn
