--- internal/pkg/service/common/etcdop/op/error.go
@@ -1,15 +1,0 @@
-package op
-
-// EmptyResultError signals an empty result of the etcd operation.
-// See also WithResult.WithEmptyResultAsError method.
-type EmptyResultError struct {
-	error
-}
-
-func (v EmptyResultError) Unwrap() error {
-	return v.error
-}
-
-func NewEmptyResultError(err error) EmptyResultError {
-	return EmptyResultError{error: err}
-}
