+++ internal/pkg/service/common/etcdop/op/op_error.go
@@ -1,0 +1,17 @@
+package op
+
+import "context"
+
+type errorOp struct {
+	err error
+}
+
+// ErrorOp wraps a static error as Op.
+// It makes error handling easier and move it to one place.
+func ErrorOp(err error) Op {
+	return &errorOp{err: err}
+}
+
+func (v *errorOp) Op(context.Context) (LowLevelOp, error) {
+	return LowLevelOp{}, v.err
+}
