+++ internal/pkg/service/common/etcdop/op/op_result.go
@@ -10,6 +10,20 @@ type Result[R any] struct {
 	result   *R
 }
 
+// EmptyResultError signals an empty result of the etcd operation.
+// See also WithResult.WithEmptyResultAsError method.
+type EmptyResultError struct {
+	error
+}
+
+func (v EmptyResultError) Unwrap() error {
+	return v.error
+}
+
+func NewEmptyResultError(err error) EmptyResultError {
+	return EmptyResultError{error: err}
+}
+
 func newResult[R any](r *RawResponse, result *R) *Result[R] {
 	return &Result[R]{
 		errors:   errors.NewMultiError(),
